+++ src/interpreter.c
@@ -212,7 +212,7 @@ static struct umbra_interpreter* umbra_interpreter(struct umbra_basic_block cons
 
     int n = 0;
 #define emit(...) p->inst[n] = (struct sw_inst){ __VA_ARGS__ }
-#define RESOLVE_PTR(inst) ((inst)->ptr.deref ? deref_slot[(int)(inst)->ptr.ix] : (inst)->ptr.bits)
+#define RESOLVE_PTR(inst) ((inst)->ptr.deref ? deref_slot[(inst)->ptr.ix] : (inst)->ptr.bits)
     for (int pass = 0; pass < 2; pass++) {
         int const lo = pass ? bb->preamble : 0, hi = pass ? bb->insts : bb->preamble;
         if (pass) { p->preamble = n; }
